// src/services/dashboard/type.ts
export interface IBaseName {
  code: string;
  name: string;
}

export interface IBaseCode extends IBaseID {
  code: string;
}

export interface IBaseLabel {
  label: string;
  value?: string;
}

export interface IBaseID {
  id: number;
  name: string;
}

export interface IAuditorium {
  code: number;
  name: string;
  auditoriumType: IBaseName;
  building: IBaseID;
}

export interface IGroup extends IBaseID {
  educationLang: IBaseName;
}

export interface IOrganization extends IBaseName {
  id: number;
  structureType: IBaseName;
  localityType: IBaseName;
  parent: null | number;
  active: boolean;
}

export interface ISubject extends IBaseName {
  id: number;
}

export interface IFile {
  url: string;
  name: string;
  size: number;
}

export interface ISemester extends IBaseName {
  id: number;
  current: boolean;
  education_year: EducationYear;
}

export interface EducationYear extends IBaseName {
  current: boolean;
}

export interface ILessonPair extends IBaseName {
  start_time: string;
  end_time: string;
}
// ...
export interface ISubjectList {
  id: number;
  _semester: string;
  grades: IGrade[];
  curriculumSubject: ICurriculumSubject;
  overallScore: IOverallScore;
  gradesByExam: IOverallScore[];
}

export interface ISubjects {
  subject: {
    id: number;
    code: string;
    name: string;
  };
  subjectType: {
    code: string;
    name: string;
  };
  _semester: number;
  total_acload: number;
  credit: number;
}

export interface IGrade {
  id: number;
  _student: number;
  semester: IBaseName;
  _subject_schedule: number;
  subject: IBaseCode;
  trainingType: IBaseName;
  employee: IBaseID;
  lessonPair: ILessonPair;
  grade: number;
  lesson_date: number;
  created_at: number;
}
// ...
export interface ICurriculumSubject {
  subject: ISubject;
  subjectType: IBaseName;
  _semester: string;
  total_acload: number;
  credit: number;
}

export interface IOverallScore {
  grade: number;
  max_ball: number;
  percent: number;
  label?: string;
  examType?: IBaseName;
}
// ...
export interface IMarkingSystem {
  code: MarkingSystemCodeEnum;
  name: string;
  count_final_exams: number;
  minimum_limit: number;
  gpa_limit: number;
  updated_at: number;
}

export enum MarkingSystemCodeEnum {
  Rating = '11',
  Grade = '12',
  Credit = '13',
}